requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
bcrypt==5.0.0
certifi==2025.11.12
cffi==2.0.0
click==8.2.1
colorama==0.4.6
cryptography==46.0.3
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.110.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.10
iniconfig==2.3.0
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
pytest==7.4.3
pytest-asyncio==0.22.0
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.36.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.29.0
